+++ ros_ulysse/changelog.md
@@ -10,3 +10,8 @@
 * **cfg/analyzers.yaml** added for _rqt_robot_monitor_
 * **LOGS** created
 * **src/celerity.py** repaired and added
+
+07/10/2019-11/10/2019
+-------------------------
+* **src/R2Sonic_acquisition.py** added
+* **LOGS/R2SONIC** created